Restructured theme commands and added theme delete command
